liste löschen und mitglieder löschen implementiert, inkl. adminfkt.: -nur admin kann mitglieder löschen -nur admin kann liste löschen, anere löschen sich selbst aus der liste; admin ist der listenersteller

+++ app/src/main/java/com/example/werkws18_21/einkaufsliste/GruppenManager.java
@@ -1,12 +1,17 @@
 package com.example.werkws18_21.einkaufsliste;
 
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.support.annotation.NonNull;
+import android.support.v7.app.AlertDialog;
 import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
+import android.view.KeyEvent;
 import android.view.View;
 import android.view.ViewGroup;
+import android.view.inputmethod.EditorInfo;
 import android.widget.Adapter;
+import android.widget.AdapterView;
 import android.widget.ArrayAdapter;
 import android.widget.Button;
 import android.widget.EditText;
@@ -17,6 +22,7 @@ import android.widget.Toast;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
+import com.google.common.util.concurrent.ListeningExecutorService;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseAuthUserCollisionException;
 import com.google.firebase.firestore.CollectionReference;
@@ -45,14 +51,18 @@ public class GruppenManager extends AppCompatActivity {
     String ListenName;
     String ListeRefString;
     ArrayList<String> Mitglieder;
+    ListView lvMitglieder;
+    private boolean owner = false;
 
     private static final String LISTEN_REFERENZ = "Listen-Referenz";
     private static final String LISTEN_NAME = "ListenName";
     private static final String LISTEN_COLLECTION = "Listen";
     private static final String USER_EMAIL = "User-Email";
     private static final String MITGLIEDER = "Mitglieder";
+    private static final String USER_ID = "User-Id";
 
     ArrayAdapter<String> adapter;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -62,14 +72,14 @@ public class GruppenManager extends AppCompatActivity {
         eTNeueEmail = findViewById(R.id.eTEmail);
         tvListenName = findViewById(R.id.tvListenName);
 
-        ListView lvMitglieder = findViewById(R.id.lvMitglieder);
+        lvMitglieder = findViewById(R.id.lvMitglieder);
         Mitglieder = new ArrayList<>();
-        adapter = new ArrayAdapter<String>(getBaseContext(), android.R.layout.simple_spinner_item, Mitglieder){
+        adapter = new ArrayAdapter<String>(getBaseContext(), android.R.layout.simple_spinner_item, Mitglieder) {
             @Override
             public View getView(int position, View convertView, ViewGroup parent) {
 
                 View view = super.getView(position, convertView, parent);
-                TextView textView5=(TextView) view.findViewById(android.R.id.text1);
+                TextView textView5 = (TextView) view.findViewById(android.R.id.text1);
                 textView5.setTextSize(20);
                 ViewGroup.LayoutParams layoutparams = view.getLayoutParams();
 
@@ -80,7 +90,8 @@ public class GruppenManager extends AppCompatActivity {
 
                 return view;
             }
-        };;
+        };
+        ;
         lvMitglieder.setAdapter(adapter);
 
         tv = findViewById(R.id.tv);
@@ -92,6 +103,64 @@ public class GruppenManager extends AppCompatActivity {
         getListenName();
         getMitglieder();
 
+        lvMitglieder.setOnItemLongClickListener(new AdapterView.OnItemLongClickListener() {
+            @Override
+            public boolean onItemLongClick(AdapterView<?> parent, View view, int position, long id) {
+
+                deleteMitglied(position);
+                return true;
+            }
+        });
+
+
+    }//onCreate-Ende
+
+    private void deleteMitglied(final int position) {
+        final CollectionReference mitglieder = ListeRef.collection(MITGLIEDER);
+        mitglieder.get().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
+            @Override
+            public void onComplete(@NonNull Task<QuerySnapshot> task) {
+                if (task.isSuccessful()) {
+                    for (QueryDocumentSnapshot doc : task.getResult()) {
+                        if (mAuth.getCurrentUser().getEmail().equals(Mitglieder.get(position))) {
+                            AlertDialog.Builder adb = new AlertDialog.Builder(GruppenManager.this);
+                            adb.setTitle("Entfernen");
+                            adb.setMessage("Sie sind der Adminsrator.\nSie können sich nicht selbst entfernen.\nLöschen Sie ggf. die Gruppe.");
+                            adb.setNegativeButton("Zurück", null);
+                            adb.show();
+                        }
+
+                        if (doc.get(USER_ID).toString().equals(mAuth.getUid().toString()) && !mAuth.getCurrentUser().getEmail().equals(Mitglieder.get(position))) {
+                            AlertDialog.Builder adb = new AlertDialog.Builder(GruppenManager.this);
+                            adb.setTitle("Entfernen?");
+                            adb.setMessage("Möchten Sie das Mitglied wirklich entfernen?");
+                            adb.setNegativeButton("Zurück", null);
+                            adb.setPositiveButton("Ok", new AlertDialog.OnClickListener() {
+                                public void onClick(DialogInterface dialog, int which) {
+                                    Query query = mitglieder.whereEqualTo(USER_EMAIL, Mitglieder.get(position));
+                                    query.get().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
+                                        @Override
+                                        public void onComplete(@NonNull Task<QuerySnapshot> task) {
+                                            if (task.isSuccessful()) {
+                                                for (QueryDocumentSnapshot doc : task.getResult()) {
+                                                    mitglieder.document(doc.getId()).delete();
+                                                }
+                                            }
+                                        }
+                                    }).addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
+                                        @Override
+                                        public void onComplete(@NonNull Task<QuerySnapshot> task) {
+                                            getMitglieder();
+                                        }
+                                    });
+                                }
+                            });
+                            adb.show();
+                        }
+                    }
+                }
+            }
+        });
     }
 
     private void getMitglieder() {
@@ -103,7 +172,7 @@ public class GruppenManager extends AppCompatActivity {
                     for (QueryDocumentSnapshot doc : task.getResult()) {
                         Mitglieder.add(doc.get(USER_EMAIL).toString());
                     }
-                    Collections.sort(Mitglieder,String.CASE_INSENSITIVE_ORDER);
+                    Collections.sort(Mitglieder, String.CASE_INSENSITIVE_ORDER);
                     adapter.notifyDataSetChanged();
                 }
             }
@@ -135,6 +204,7 @@ public class GruppenManager extends AppCompatActivity {
             return;
         }
         Map<String, Object> UserMap = new HashMap<>();
+        UserMap.put(USER_ID, "");
         UserMap.put(USER_EMAIL, sEmail);
         ListeRef.collection(MITGLIEDER).document().set(UserMap).addOnSuccessListener(new OnSuccessListener<Void>() {
             @Override
@@ -148,13 +218,61 @@ public class GruppenManager extends AppCompatActivity {
         toListe(ListeRefString);
     }
 
-    public void deleteListe(View view){
-        ListeRef.delete().addOnCompleteListener(new OnCompleteListener<Void>() {
-            @Override
-            public void onComplete(@NonNull Task<Void> task) {
-                toGruppenauswahl();
+    public void deleteListe(View view) {
+        final CollectionReference mitglieder = ListeRef.collection(MITGLIEDER);
+
+        AlertDialog.Builder adb = new AlertDialog.Builder(GruppenManager.this);
+        adb.setTitle("Entfernen?");
+        adb.setMessage("Möchten Sie die Liste wirklich löschen?");
+        adb.setNegativeButton("Zurück", null);
+        adb.setPositiveButton("Ok", new AlertDialog.OnClickListener() {
+            public void onClick(DialogInterface dialog, int which) {
+                mitglieder.get().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
+                    @Override
+                    public void onComplete(@NonNull Task<QuerySnapshot> task) {
+                        if (task.isSuccessful()) {
+                            for (QueryDocumentSnapshot doc : task.getResult()) {
+                                if (mAuth.getUid().equals(doc.get(USER_ID).toString())) {
+                                    owner = true;
+                                    ListeRef.delete().addOnCompleteListener(new OnCompleteListener<Void>() {
+                                        @Override
+                                        public void onComplete(@NonNull Task<Void> task) {
+                                            toGruppenauswahl();
+                                        }
+                                    });
+                                }
+                            }
+                        }
+                    }
+                }).addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
+                    @Override
+                    public void onComplete(@NonNull Task<QuerySnapshot> task) {
+                        if (owner == false) {
+                            Query query = mitglieder.whereEqualTo(USER_EMAIL, mAuth.getCurrentUser().getEmail());
+                            query.get().addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
+                                @Override
+                                public void onComplete(@NonNull Task<QuerySnapshot> task) {
+                                    if (task.isSuccessful()) {
+                                        for (QueryDocumentSnapshot doc : task.getResult()) {
+                                            mitglieder.document(doc.getId()).delete().addOnCompleteListener(new OnCompleteListener<Void>() {
+                                                @Override
+                                                public void onComplete(@NonNull Task<Void> task) {
+                                                    toGruppenauswahl();
+                                                }
+                                            });
+                                        }
+                                    }
+                                }
+                            });
+                        }
+                    }
+                });
             }
         });
+
+        adb.show();
+
+
     }
 
     private void toGruppenauswahl() {
